lag TrygdeytelserInputs komponent
